+++ src/components/FamilyDiagram.tsx
@@ -2,7 +2,7 @@ import { Person, PersonNode } from '../family.interface';
 import { treesToPersonNode } from '../utils/family-tree';
 import { useRef, useEffect, useDeferredValue, useState } from 'react';
 import * as go from 'gojs';
-import { isEqual, cloneDeep } from 'lodash';
+import { isEqual, cloneDeep } from 'lodash-es';
 import GenogramLayout from '../GenogramLayout';
 
 const SLASH_COLOR = '#d4071c'; // red